Add copyright statement. Add support for opening in read only mode.

+++ index/src/TopdocsIndex.cpp
@@ -1,3 +1,13 @@
+/*==========================================================================
+ * Copyright (c) 2004 University of Massachusetts.  All Rights Reserved.
+ *
+ * Use of the Lemur Toolkit for Language Modeling and Information Retrieval
+ * is subject to the terms of the software license set forth in the LICENSE
+ * file included with this software, and also available at
+ * http://www.lemurproject.org/license.html
+ *
+ *==========================================================================
+*/
 
 //
 // TopdocsIndex
@@ -17,7 +27,7 @@
 void TopdocsIndex::create( const std::string& pathname ) {
   _documents = 0;
   _pathname = pathname;
-
+  _readOnly = false;
   // makedir filename
   if( Path::exists( _pathname ) )
     Path::remove( _pathname );
@@ -30,6 +40,7 @@ void TopdocsIndex::create( const std::string& pathname ) {
 }
 
 void TopdocsIndex::openRead( const std::string& pathname ) {
+  _readOnly = true;
   _pathname = pathname;
   Parameters param;
 
@@ -44,6 +55,7 @@ void TopdocsIndex::openRead( const std::string& pathname ) {
 
 void TopdocsIndex::open( const std::string& pathname ) {
   _pathname = pathname;
+  _readOnly = false;
   Parameters param;
   
   std::string manifestPath = Path::combine( _pathname, "manifest" );
@@ -55,11 +67,12 @@ void TopdocsIndex::open( const std::string& pathname ) {
 }
 
 void TopdocsIndex::close() {
-  Parameters param;
-  param.set( "documents", _documents );
-  std::string manifestPath = Path::combine( _pathname, "manifest" );
-
-  param.writeFile( manifestPath );
+  if (! _readOnly) {
+    Parameters param;
+    param.set( "documents", _documents );
+    std::string manifestPath = Path::combine( _pathname, "manifest" );
+    param.writeFile( manifestPath );
+  }
   _lists.close();
 }
 
@@ -188,6 +201,8 @@ TopdocsIndex::TopdocsList* TopdocsIndex::fetch( int term ) {
       output->entries.push_back( entry );
     }
 
+    // delete the buffer allocated by get
+    delete[](data);
     return output;
   } else {
     return 0;
